release-22.1: sql: fix overflow for soft limits

+++ pkg/sql/distsql_physical_planner.go
@@ -2250,7 +2250,7 @@ func (dsp *DistSQLPlanner) createPlanForIndexJoin(
 		LockingStrength:   n.table.lockingStrength,
 		LockingWaitPolicy: n.table.lockingWaitPolicy,
 		MaintainOrdering:  len(n.reqOrdering) > 0,
-		LimitHint:         int64(n.limitHint),
+		LimitHint:         n.limitHint,
 	}
 
 	fetchColIDs := make([]descpb.ColumnID, len(n.cols))
@@ -2319,7 +2319,7 @@ func (dsp *DistSQLPlanner) createPlanForLookupJoin(
 		LeftJoinWithPairedJoiner:          n.isSecondJoinInPairedJoiner,
 		OutputGroupContinuationForLeftRow: n.isFirstJoinInPairedJoiner,
 		LookupBatchBytesLimit:             dsp.distSQLSrv.TestingKnobs.JoinReaderBatchBytesLimit,
-		LimitHint:                         int64(n.limitHint),
+		LimitHint:                         n.limitHint,
 	}
 
 	fetchColIDs := make([]descpb.ColumnID, len(n.table.cols))
